ARM64 / ACPI: Implement acpi_map/unmap_lsapic for cpu hot-plug
When cpu was hot-added or hot removed, the cpu_present_map and APIC id map
to logic cpu id should be updated. acpi_map/unmap_lsapic() will do this.

available_cpus will be used for ACPI driver, so remove __initdata attribute
for it, and change it into global for future use too.

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>