ARM64 / CPU hot-plug: Skeleton logic cpu online/offline for cpu hot-plug
CPU logic online/offline was not implemented on arm64 which is essential for
cpu hot-plug, so skeleton the logic cpu online/offline to make arm64 code
compile when CONFIG_HOTPLUG_CPU was introduced.

It should be fully implemented at later date. Will discuss it with power
management team in LCE13.

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>