Add system table pointer argument to shared functions.
Add system table pointer argument to shared EFI stub related functions
so the no longer use global system table pointer as they did when part
of eboot.c.  For the ARM EFI stub this allows us to avoid global
variables completely and thereby not have to deal with GOT fixups.
This code is now shared, so using a global variable as part of the
interface is not that nice. Also, by avoiding any global variables in
the ARM EFI stub, this allows the code to be position independent
without requiring GOT fixups.  Not having the EFI stub fixup its GOT,
which is shared with the decompressor, simplifies the relocating
of the zImage to a bootable address.


Signed-off-by: Roy Franz <roy.franz@linaro.org>