ACPI: Add support for device specific properties
Device Tree is used in many embedded systems to describe the system
configuration to the OS. It supports attaching properties or name-value
pairs to the devices it describe. With these properties one can pass
additional information to the drivers that would not be available
otherwise.

ACPI is another configuration mechanism (among other things) typically
seen on but not limited to x86 machines. ACPI allows passing arbitrary
data from methods but there is nothing like Device Tree properties in
the ACPI specification.

In order to facilitate ACPI usage in systems where Device Tree is
typically used, it would be beneficial to standardize a way to retrieve
Device Tree style properties from ACPI devices, which is what we do in
this patch.

If a given device described in ACPI namespace wants to export properties
it must implement _PRP method that returns the properties in a package
of packages. For example:

	Method (_PRP, 0, NotSerialized)
	{
		Return (Package () {
			Package () { "name1", <VALUE1> },
			Package () { "name2", <VALUE2> },
			...
		})
	}

Name must be string but there are no limitations for the value itself.

We add several helper functions that can be used to extract these
properties and convert them to different Linux data types.

The ultimate goal is that we only have one device property API that
retrieves the requested properties from Device Tree or from ACPI
transparent to the caller.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>