gpio_keys_polled: Add ACPI enumeration and MinnowBoard GPIO button support
Add support for the MinnowBoard GPIO buttons described by the ACPI _HID
MNW0002.

This is an experimental mechanism to extract GPIO pins and keycodes from
a custom SSDT. It extracts the pins/buttons from the first GpioIO
resource pin_table. It extracts keycodes from the vendor_data, which
must contain a matching number of keycodes.

There are a number of platform data fields for the gpio_keys* driver
that are not readily implemented with ACPI 5, for these, default values
are assigned.

This mechanism is not idea as it defines how the vendor data must look
for any device looking to use gpio_keys_polled. It also encodes
Linux-specific keycode values into an OS agnostic firmware construct.

A better approach might be to move the pdata generation from ACPI into
its own drivers/platform/x86/ driver which then loads the platform
driver, thus leaving the gpio_keys_polled driver independent from the
ACPI description method. This doesn't scale well, but it would allow
time for the ACPI specification to mature in terms of generic hardware
description before a generic format is defined in the gpio_keys*
drivers.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>