ACPI/ARM: Load fixed-clk module early
An AMBA device will be probed once two things have occured: 1) ACPI has
registered a device, and 2) the driver has registered itself as an AMBA
driver. Since several drivers register themselves very early on, the
probe will happen as soon as ACPI registers the device. If a device
depends on a clock, the clock must be probed before ACPI registers the
device. This means that the clock definition must be encountered first in the
DSDT file, and the clock driver must already be loaded when this happens.

There are other potential solutions, with this solution registering the clock
driver very early in the boot process to avoid changes to other drivers.

Signed-off-by: Brandon Anderson <brandon.anderson@amd.com>
Reviewed-by: Naresh Bhat <naresh.bhat@linaro.org>