ACPI: GUFI: add in early prototype code for some GUFI functions
This patch introduces the basic structure and content for just a couple
of example GUFI functions.  Each function is designed so that one can
continue to use FDT and with minimal change, convert to using functions
that will retrieve configuration data from either DT or ACPI, whichever
has info available.  In essence, these functions act as a translation
layer.

Signed-off-by: Al Stone <al.stone@linaro.org>