ACPI: GUFI: start modifying the vexpress-sysreg driver to use the GUFI
As a way to test out the concepts and implementation, start converting the
existing vexpress-sysreg driver so that it uses the GUFI.  The intent is to
(a) act as a test for the idea, and (b) be very clear on how extensive (or
not) the changes actually are.  The goal is to make it so the changes can
be done almost entirely with a sed script.

Signed-off-by: Al Stone <al.stone@linaro.org>