ACPI: GUFI: temporarily use the method "PROP" instead of "_PRP"
In ASL, the name "_PRP" would have to be a predefined name because of the
leading underscore.  If it is a predefined name, we have to change iasl to
recognize it as such so we can build tables.  Getting that change done is
a much longer task; changing the name to "PROP" for the short term allows
us to use iasl as is and continue experimenting with the idea.

Signed-off-by: Al Stone <al.stone@linaro.org>