ARM64 / ACPI: Get the enable method for SMP initialization
ACPI has no flag to indicate different enable methods for SMP
initialization, but it indicates that spin-table is supported
at now because of the Parked Protocol shows that there is
parked address in GIC structure.

In order to boot the system with ACPI if DT is not available,
we set the default enable method as spin-table since PSCI is
not available in ACPI spec at now.

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>