ACPI: introduce CONFIG_ACPI_REDUCED_HARDWARE_ONLY to enable this ACPI mode
ACPI hardware reduced mode exists to allow newer platforms to use a
simpler form of ACPI that does not require supporting legacy versions
of the specification and their associated hardware.  This mode was
introduced in the ACPI 5.0 specification.

To enable the hardware reduced mode of ACPI, we need to set the flag
ACPI_REDUCED_HARDWARE to TRUE in the ACPICA source.  In order to do
that, we introduce a kernel configuration item to enable or disable
ACPI_REDUCED_HARDWARE.  We can then change the kernel configuration
instead of having to modify the kernel source.

Introducing this configuration item is based on suggestions from Lv
Zheng saying that this does not belong in ACPICA, but rather to the
Linux kernel itself.  Hence, we introduce this configuration item so
that we can make ACPI_REDUCED_HARDWARE configurable.  For the details
of the discussion, please refer to:

   http://www.spinics.net/lists/linux-acpi/msg46369.html

Changes for v3:
   -- Minimize the changelog.

Changes for v2:
   -- Changed test for EXPERT to avoid reported Kconfig warning

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>
Signed-off-by: Al Stone <al.stone@linaro.org>