GUFI: correct gufi_look_for_node() so it updates the data it searches
The gufi_look_for_node() functions searchs a list of gufi_device_nodes
that each have a pointer to a DT and ACPI node.  Make sure those nodes
reflect the latest search results so we don't end up with a list of
GUFI nodes that have either ACPI pointers and null DT pointers, or DT
pointers with null ACPI pointers.

Signed-off-by: Al Stone <al.stone@linaro.org>