arm64: vexpress: HACK: add "dcc" and "position" to device tree
This patch adds "dcc" and "position" properties to VE device tree file.
Actually those properties can be ignored in device tree file, because
they are "0" and driver code can set them "0" if they are not defined
in device tree (binding documentation says that non existing property means 0).

So this patch only needed for testing "sysreg" driver when it will be
ported to GUFI framework. After that this patch can be reverted.

Signed-off-by: Semen Protsenko <semen.protsenko@linaro.org>